update: constants types

+++ src/jam_params.zig
@@ -10,11 +10,11 @@ pub const Params = struct {
     // A: The period, in seconds, between audit tranches
     audit_tranche_period: u8 = 8, // A
     // BI: The additional minimum balance required per item of elective service state
-    min_balance_per_item: u32 = 10, // BI
+    min_balance_per_item: u64 = 10, // BI
     // BL: The additional minimum balance required per octet of elective service state
-    min_balance_per_octet: u32 = 1, // BL
+    min_balance_per_octet: u64 = 1, // BL
     // BS: The basic minimum balance which all services require
-    basic_service_balance: u32 = 100, // BS
+    basic_service_balance: u64 = 100, // BS
     // C: The total number of cores
     core_count: u16 = 341, // C
     // D: The period in timeslots after which an unreferenced preimage may be expunged
@@ -25,40 +25,40 @@ pub const Params = struct {
     audit_bias_factor: u8 = 2, // F
     // GA: The total gas allocated to a core for Accumulation
     // See: https://github.com/w3f/jamtestvectors/pull/20#issuecomment-2526203035
-    gas_alloc_accumulation: u32 = 10_000_000, // GA
+    gas_alloc_accumulation: u64 = 10_000_000, // GA
     // GI: The gas allocated to invoke a work-package Is-Authorized logic
-    gas_alloc_is_authorized: u32 = 50_000_000, // GI
+    gas_alloc_is_authorized: u64 = 50_000_000, // GI
     // GR: The total gas allocated for a work-package Refine logic
     gas_alloc_refine: u64 = 5_000_000_000, // GR
     // GT: The total gas allocated across all cores for Accumulation
     total_gas_alloc_accumulation: u64 = 3_500_000_000, // GT
     // H: The size of recent history, in blocks
-    recent_history_size: u8 = 8, // H
+    recent_history_size: u16 = 8, // H
     // I: The maximum amount of work items in a package
-    max_work_items_per_package: u8 = 16, // I
+    max_work_items_per_package: u16 = 16, // I
     // J: The maximum amount of dependencies in a work report segment-root
     // lookup dictionary and the number of pre-requisites for a work item
-    max_number_of_dependencies_for_work_reports: u8 = 8, // J
+    max_number_of_dependencies_for_work_reports: u16 = 8, // J
     // K: The maximum number of tickets which may be submitted in a single extrinsic
     max_tickets_per_extrinsic: u32 = 16, // K
     // L: The maximum age in timeslots of the lookup anchor
     max_lookup_anchor_age: u32 = 14_400, // L
     // N: The number of ticket entries per validator
-    max_ticket_entries_per_validator: u8 = 2, // N
+    max_ticket_entries_per_validator: u16 = 2, // N
     // O: The maximum number of items in the authorizations pool
-    max_authorizations_pool_items: u8 = 8, // O // TODO: rename to authorizors
+    max_authorizations_pool_items: u16 = 8, // O // TODO: rename to authorizors
     // P: The slot period, in seconds
-    slot_period: u8 = 6, // P TODO: integrate these params into Tau and Safrole
+    slot_period: u16 = 6, // P TODO: integrate these params into Tau and Safrole
     // Q: The maximum number of items in the authorizations queue
-    max_authorizations_queue_items: u8 = 80, // Q
+    max_authorizations_queue_items: u16 = 80, // Q
     // R: The rotation period of validator-core assignments, in timeslots
     validator_rotation_period: u32 = 10, // R
     // S: The maximum number of entries in the accumulation queue
     max_accumulation_queue_entries: u32 = 1024, // S
     // T: The maximum number of extrinsics in a work-package
-    max_extrinsics_per_work_package: u8 = 128, // T
+    max_extrinsics_per_work_package: u16 = 128, // T
     // U: The period in timeslots after which reported but unavailable work may be replaced
-    work_replacement_period: u8 = 5, // U
+    work_replacement_period: u16 = 5, // U
     // V: The total number of validators
     validators_count: u32 = 1023, // V
     // V_s: The number of validators required for a super-majority
@@ -70,19 +70,19 @@ pub const Params = struct {
     // WC: The maximum size of service code in octets
     max_service_code_size: u32 = 4_000_000, // WC
     // WE: The basic size of erasure-coded pieces in octets
-    erasure_coded_piece_size: u16 = 684, // WE
+    erasure_coded_piece_size: u32 = 684, // WE
     // WG: The size of a segment in octets (WP * WE)
     segment_size: u16 = 4104, // WG
     // WM: The maximum number of imports in a work-package
     max_imports_per_work_package: u16 = 3072, // WM
     // WP: The number of erasure-coded pieces in a segment
-    erasure_coded_pieces_per_segment: u8 = 6, // WP
+    erasure_coded_pieces_per_segment: u32 = 6, // WP
     // WR: The maximum size of an encoded work-report in octets
     max_work_report_size: u32 = 48 * (1 << 10), // WR
     // WS: The size of an exported segment in erasure-coded pieces (same as WP)
-    exported_segment_size: u8 = 6, // WS
+    exported_segment_size: u32 = 6, // WS
     // WT: The size of a transfer memo in octets
-    transfer_memo_size: u8 = 128, // WT
+    transfer_memo_size: u32 = 128, // WT
     // WX: The maximum number of exports in a work-package
     max_exports_per_work_package: u16 = 3072, // WX
     // Y: The number of slots into an epoch at which ticket-submission ends
@@ -101,7 +101,7 @@ pub const Params = struct {
     // -- (cores-count + 7) / 8
     avail_bitfield_bytes: usize = (341 + 7) / 8,
 
-    pub fn segmentSizeInOctets(comptime self: *const Params) u16 {
+    pub fn segmentSizeInOctets(comptime self: *const Params) u32 {
         // WG = WP * WE
         return self.erasure_coded_pieces_per_segment * self.erasure_coded_piece_size;
     }
